pyproject.toml
[tool.poetry]
name = "docspider-documents-backend"
version = "0.1.0"
description = ""
authors = ["Bruno Werner <brunowerner7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
prettyconf = "^2.2.1"
django-jazzmin = "^3.0.0"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
django-prometheus = "^2.3.1"
django-extensions = "^3.2.3"
django-rest-passwordreset = "^1.5.0"
drf-spectacular = "^0.28.0"
django-cors-headers = "^4.6.0"
django-filter = "^24.3"
six = "^1.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
